@@ -3,8 +3,9 @@
   "version": "2.0.0",
   "description": "Based on Single Page Apps for GitHub Pages from rafrex/spa-github-pages.git",
   "scripts": {
-    "build": "webpack -p",
-    "start": "webpack -d --watch & live-server --entry-file=404.html"
+    "build": "webpack --mode production",
+    "start": "webpack --mode development --watch & live-server --entry-file=404.html",
+    "lint": "eslint --ext .jsx,.js ."
   },
   "repository": {
     "type": "git",
@@ -13,33 +14,33 @@
   "author": "Jared Casner <jared@casner.us>",
   "license": "MIT",
   "dependencies": {
-    "moment": "2.24.0",
-    "prop-types": "15.7.2",
-    "query-string": "6.10.1",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-router-dom": "5.1.2"
+    "moment-mini": "2.24.0",
+    "prop-types": "15.8.1",
+    "query-string": "7.1.1",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-router-dom": "6.3.0"
   },
   "devDependencies": {
-    "@babel/core": "7.8.3",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "babel-eslint": "10.0.3",
-    "babel-loader": "8.0.6",
-    "eslint": "6.8.0",
-    "eslint-config-airbnb": "18.0.1",
-    "eslint-plugin-import": "2.20.0",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "^6.0.0",
+    "@babel/core": "7.17.9",
+    "@babel/eslint-parser": "7.17.0",
+    "@babel/preset-env": "7.16.11",
+    "@babel/preset-react": "7.16.7",
+    "@babel/register": "^7.17.7",
+    "babel-loader": "8.2.4",
+    "eslint": "^8.13.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "6.5.1",
+    "eslint-plugin-react": "7.29.4",
+    "eslint-plugin-react-hooks": "4.4.0",
     "live-server": "1.2.1",
-    "terser-webpack-plugin": "^2.3.5",
-    "webpack": "4.42.1",
-    "webpack-cli": "3.3.11"
+    "terser-webpack-plugin": "5.3.1",
+    "webpack": "5.72.0",
+    "webpack-cli": "4.9.2"
   },
   "resolutions": {
-    "acorn": "7.4.1",
-    "ansi-regex": "^5.0.1",
-    "glob-parent": "^5.1.2",
-    "set-value": "^4.0.1"
+    "fsevents": "2.3.2",
+    "glob-parent": "5.1.2"
   }
 }
