@@ -2,7 +2,7 @@ import React from 'react';
 import {
   BrowserRouter as Router,
   Route,
-  Switch
+  Routes
 } from 'react-router-dom';
 
 import Home from './home';
@@ -10,11 +10,11 @@ import Resume from './resume';
 import PageNotFound from './not-found';
 
 const routes = (
-  <Switch>
-    <Route exact path="/" title="Home" component={Home} />
-    <Route path="/resume" title="Resume" component={Resume} />
-    <Route component={PageNotFound} />
-  </Switch>
+  <Routes>
+    <Route exact path="/" title="Home" element={<Home />} />
+    <Route path="/resume" title="Resume" element={<Resume />} />
+    <Route element={<PageNotFound />} />
+  </Routes>
 );
 
 const App = () => {
